+++ Code/Editor/ConfigGroup.cpp
@@ -33,7 +33,7 @@ namespace Config
 
     uint32 CConfigGroup::GetVarCount()
     {
-        return static_cast<uint32>(m_vars.size());
+        return m_vars.size();
     }
 
     IConfigVar* CConfigGroup::GetVar(const char* szName)
