+++ Gems/Atom/RHI/Code/Source/RHI/FrameGraph.cpp
@@ -497,7 +497,7 @@ namespace AZ
                 for (const uint32_t edgeIndex : graphEdges[producerIndex])
                 {
                     const GraphEdge& graphEdge = m_graphEdges[edgeIndex];
-                    const uint16_t consumerIndex = static_cast<uint16_t>(graphEdge.m_consumerIndex);
+                    const uint16_t consumerIndex = graphEdge.m_consumerIndex;
                     if (--m_graphNodes[consumerIndex].m_unsortedProducerCount == 0)
                     {
                         NodeId newNode;
