+++ Gems/Atom/RHI/Vulkan/Code/Source/RHI/Conversion.cpp
@@ -257,7 +257,7 @@ namespace AZ
             state.srcAlphaBlendFactor = ConvertBlendFactor(targetBlendState.m_blendAlphaSource);
             state.dstAlphaBlendFactor = ConvertBlendFactor(targetBlendState.m_blendAlphaDest);
             state.alphaBlendOp = ConvertBlendOp(targetBlendState.m_blendAlphaOp);
-            state.colorWriteMask = ConvertComponentFlags(static_cast<uint8_t>(targetBlendState.m_writeMask));
+            state.colorWriteMask = ConvertComponentFlags(targetBlendState.m_writeMask);
         }
 
         VkBlendFactor ConvertBlendFactor(const RHI::BlendFactor& blendFactor)
