+++ Gems/AtomLyIntegration/CommonFeatures/Code/Source/CoreLights/DiskLightDelegate.cpp
@@ -159,7 +159,7 @@ namespace AZ::Render
     {
         if (GetShadowsEnabled() && GetLightHandle().IsValid())
         {
-            GetFeatureProcessor()->SetPredictionSampleCount(GetLightHandle(), static_cast<uint16_t>(count));
+            GetFeatureProcessor()->SetPredictionSampleCount(GetLightHandle(), count);
         }
     }
 
@@ -167,7 +167,7 @@ namespace AZ::Render
     {
         if (GetShadowsEnabled() && GetLightHandle().IsValid())
         {
-            GetFeatureProcessor()->SetFilteringSampleCount(GetLightHandle(), static_cast<uint16_t>(count));
+            GetFeatureProcessor()->SetFilteringSampleCount(GetLightHandle(), count);
         }
     }
