+++ Gems/EMotionFX/Code/Tests/AnimGraphMotionNodeTests.cpp
@@ -303,7 +303,7 @@ namespace EMotionFX
 
     TEST_F(AnimGraphMotionNodeFixture, InPlaceInputAndNoEffectOutputsCorrectMotionAndPose)
     {
-        m_motionNode->AddConnection(m_paramNode, static_cast<uint16>(m_paramNode->FindOutputPortByName("InPlace")->m_portId), AnimGraphMotionNode::INPUTPORT_INPLACE);
+        m_motionNode->AddConnection(m_paramNode, m_paramNode->FindOutputPortByName("InPlace")->m_portId, AnimGraphMotionNode::INPUTPORT_INPLACE);
         ParamSetValue<MCore::AttributeBool, bool>("InPlace", true);
         m_animGraphInstance->FindOrCreateUniqueNodeData(m_motionNode);
